Adds get-deployment-by-id which uses latest v9 endpoint
